HaplotypeCaller: the engine-level pieces
Included in this PR are the engine-level dependencies of the HaplotypeCaller
from the dr_runnable_haplotypecaller branch. This includes all parts
of that branch that lie outside of the haplotypecaller package itself, and
mostly consists of a new traversal (AssemblyRegionWalker), additions to
the genotyper package, and various additions/tweaks to common utilities
and engine-level classes.

The haplotypecaller package itself has been removed in preparation for its
migration to the gatk-protected repository.